requirements.txt
requests
paho-mqtt
beautifulsoup4